+++ v3/crates/engine/tests/execution.rs
@@ -3072,3 +3072,23 @@ fn test_variables_non_null_type_default_value() -> anyhow::Result<()> {
         common::TestOpenDDPipeline::YesPlease,
     )
 }
+
+#[test]
+fn test_variables_non_null_type_default_value_null() -> anyhow::Result<()> {
+    let test_path_string = "execute/variables/non_null_type_default_value_null";
+    common::test_execution_expectation_for_multiple_ndc_versions(
+        test_path_string,
+        &[],
+        BTreeMap::from([
+            (
+                NdcVersion::V01,
+                vec!["execute/common_metadata/postgres_connector_ndc_v01_schema.json"],
+            ),
+            (
+                NdcVersion::V02,
+                vec!["execute/common_metadata/postgres_connector_ndc_v02_schema.json"],
+            ),
+        ]),
+        common::TestOpenDDPipeline::YesPlease,
+    )
+}